+++ types/settings/DerivePasswordForWriteInput.ts
@@ -1,0 +1,6 @@
+export type DerivePasswordForWriteInput = {
+  readonly hasNewPassword: boolean
+  readonly newPassword: string | undefined
+  readonly existingPassword: string | undefined
+  readonly secretKey: string
+}